+++ packages/apm/src/integrations/tracing.ts
@@ -286,6 +286,44 @@ export class Tracing implements Integration {
   }
 
   /**
+   * Returns a new Transaction either continued from sentry-trace meta or a new one
+   */
+  private static _getNewTransaction(hub: Hub, transactionContext: TransactionContext): Transaction {
+    let traceId;
+    let parentSpanId;
+    let sampled;
+
+    const header = Tracing._getMeta('sentry-trace');
+    if (header) {
+      const span = SpanClass.fromTraceparent(header);
+      if (span) {
+        traceId = span.traceId;
+        parentSpanId = span.parentSpanId;
+        sampled = span.sampled;
+        Tracing._log(
+          `[Tracing] found 'sentry-meta' '<meta />' continuing trace with: trace_id: ${traceId} span_id: ${parentSpanId}`,
+        );
+      }
+    }
+
+    return hub.startTransaction({
+      parentSpanId,
+      sampled,
+      traceId,
+      trimEnd: true,
+      ...transactionContext,
+    }) as Transaction;
+  }
+
+  /**
+   * Returns the value of a meta tag
+   */
+  private static _getMeta(metaName: string): string | null {
+    const el = document.querySelector(`meta[name=${metaName}]`);
+    return el ? el.getAttribute('content') : null;
+  }
+
+  /**
    * Pings the heartbeat
    */
   private static _pingHeartbeat(): void {
@@ -456,10 +494,7 @@ export class Tracing implements Integration {
       return undefined;
     }
 
-    Tracing._activeTransaction = hub.startTransaction({
-      trimEnd: true,
-      ...transactionContext,
-    }) as Transaction;
+    Tracing._activeTransaction = Tracing._getNewTransaction(hub, transactionContext);
 
     // We set the transaction here on the scope so error events pick up the trace context and attach it to the error
     hub.configureScope(scope => scope.setSpan(Tracing._activeTransaction));