+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - [tracing] fix: Add manual Location typing (#2700)
 - [react] feat: Expose eventId on ErrorBoundary component
 
+- [tracing] feat: Pick up sentry-trace in JS <meta/> tag (#2703)
+- [tracing] fix: Respect sample decision when continuing trace from header in node (#2703)
+
 ## 5.18.1
 
 - [react] feat: Update peer dependencies for `react` and `react-dom` (#2694)